server/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
flask>=2.0.0
flask-cors>=3.0.0
pillow>=9.0.0
pillow-heif>=0.13.0
tqdm>=4.65.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
scikit-image>=0.19.0
tensorboard>=2.11.0